Currently we can acquire link robustly but we still occasionally drop it. We're using physical devices 0-1, 6-7 as logical 0-3. Not clear what the next steps are.
As of this commit we've noticed that syscontrol appears to be
double-sending some events, in particular the link drop count events.
We've added some counters and it does appear that they are being
double set at some point outside the eproc core. This might be
because of subtle clock phase issues caused by slight changes
to reset timing earlier? Not sure.